requirements.txt
fastapi
uvicorn
pydantic
requests
openai
python-dotenv
streamlit
langchain
langchain-core
langchain-community
langchain-text-splitters
faiss-cpu
tqdm
python-docx
Pillow
huggingface-hub
duckduckgo-search
docx2txt
sentence-transformers
ddgs
rank-bm25
jieba